moving commands to scene-relative files, reset assignments in scenes
