+++ 2D3D_UnityProject/Assets/Prefabs/Dialogue.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 8f0facbab6e2a9f4b8b028fb7147ebf2
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
