+++ 2D3D_UnityProject/Assets/Scripts/Yarnspinner Commands.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 19c45f0234a8b9747a2916717be9da7e
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
