+++ 2D3D_UnityProject/Assets/Scripts/Yarnspinner Commands/LiarCommands.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 7d5634f29acea4448b40e0680b79079a
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
